Reorganizing files and folder structure
