+++ bundles/framework/layerlist/view/LayerViewTabs/LayerList/LayerCollapse/CollapseService.js
