+++ bundles/framework/layerlist/view/LayerViewTabs/LayerList/LayerCollapse/Layer/index.js
@@ -1,0 +1,1 @@
+export { Layer } from './Layer';
