+++ bundles/framework/layerlist/view/LayerViewTabs/LayerList/LayerCollapse/index.js
@@ -1,0 +1,2 @@
+export { LayerCollapse } from './LayerCollapse';
+export { CollapseService, GROUPING_METHODS } from './CollapseService';
