+++ bundles/framework/layerlist/view/LayerViewTabs/LayerList/LayerFilters/index.js
@@ -1,0 +1,2 @@
+export { LayerFilters } from './LayerFilters';
+export { FilterService } from './FilterService';
