+++ bundles/framework/layerlist/view/LayerViewTabs/LayerList/index.js
@@ -1,0 +1,2 @@
+export { LayerList } from './LayerList';
+export { LayerListAlert } from './LayerListAlert';
