+++ bundles/framework/layerlist/view/LayerViewTabs/LayerViewTabs.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { LayerList } from './LayerList/';
+
+export const LayerViewTabs = props => (
+    <LayerList {...props} />
+);
