+++ bundles/framework/layerlist/view/LayerViewTabs/index.js
@@ -1,0 +1,1 @@
+export { LayerViewTabs } from './LayerViewTabs';
